# update to go1.18, and use strings.Cut