--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -4,7 +4,7 @@ jobs:
   test:
     strategy:
       matrix:
-        go-version: [1.17.x, 1.21.x, 1.22.x]
+        go-version: [1.18.x, 1.21.x, 1.22.x]
         platform: [ubuntu-20.04, ubuntu-22.04, ubuntu-24.04, windows-latest, macos-12, macos-14]
     runs-on: ${{ matrix.platform }}
     steps: