--- a/mount/go.mod
+++ b/mount/go.mod
@@ -1,6 +1,6 @@
 module github.com/moby/sys/mount
 
-go 1.17
+go 1.18
 
 require (
 	github.com/moby/sys/mountinfo v0.7.2