--- a/mount/mounter_linux_test.go
+++ b/mount/mounter_linux_test.go
@@ -212,7 +212,8 @@ func validateMount(t *testing.T, mnt string, opts, optional, vfs string) {
 
 // clean strips off any value param after the colon
 func clean(v string) string {
-	return strings.SplitN(v, ":", 2)[0]
+	out, _, _ := strings.Cut(v, ":")
+	return out
 }
 
 // has returns true if key is a member of m