--- a/mountinfo/mountinfo_linux.go
+++ b/mountinfo/mountinfo_linux.go
@@ -75,7 +75,10 @@ func GetMountsFromReader(r io.Reader, filter FilterFunc) ([]*Info, error) {
 			}
 		}
 
-		p := &Info{}
+		p := &Info{
+			ID:     toInt(fields[0]),
+			Parent: toInt(fields[1]),
+		}
 
 		p.Mountpoint, err = unescape(fields[4])
 		if err != nil {
@@ -91,15 +94,12 @@ func GetMountsFromReader(r io.Reader, filter FilterFunc) ([]*Info, error) {
 		}
 		p.VFSOptions = fields[sepIdx+3]
 
-		// ignore any numbers parsing errors, as there should not be any
-		p.ID, _ = strconv.Atoi(fields[0])
-		p.Parent, _ = strconv.Atoi(fields[1])
 		mm := strings.SplitN(fields[2], ":", 3)
 		if len(mm) != 2 {
 			return nil, fmt.Errorf("parsing '%s' failed: unexpected major:minor pair %s", text, mm)
 		}
-		p.Major, _ = strconv.Atoi(mm[0])
-		p.Minor, _ = strconv.Atoi(mm[1])
+		p.Major = toInt(mm[0])
+		p.Minor = toInt(mm[1])
 
 		p.Root, err = unescape(fields[3])
 		if err != nil {
@@ -247,4 +247,11 @@ func unescape(path string) (string, error) {
 	}
 
 	return string(buf[:bufLen]), nil
+}
+
+// toInt converts a string to an int, and ignores any numbers parsing errors,
+// as there should not be any.
+func toInt(s string) int {
+	i, _ := strconv.Atoi(s)
+	return i
 }