--- a/sequential/go.mod
+++ b/sequential/go.mod
@@ -1,5 +1,5 @@
 module github.com/moby/sys/sequential
 
-go 1.17
+go 1.18
 
 require golang.org/x/sys v0.1.0