--- a/symlink/go.mod
+++ b/symlink/go.mod
@@ -1,5 +1,5 @@
 module github.com/moby/sys/symlink
 
-go 1.17
+go 1.18
 
 require golang.org/x/sys v0.1.0