--- a/user/go.mod
+++ b/user/go.mod
@@ -1,5 +1,5 @@
 module github.com/moby/sys/user
 
-go 1.17
+go 1.18
 
 require golang.org/x/sys v0.1.0